+++ maman22/mycomp/complex.c
@@ -2,7 +2,7 @@
 #include "complex.h"
 
 /*initialize a complex number*/
-Complex getComplex(float real_part, float imaginary_part) {
+Complex getComplex(double real_part, double imaginary_part) {
     Complex c;
     c.real_part = real_part;
     c.imaginary_part = imaginary_part;
@@ -21,14 +21,14 @@ Complex sub(Complex num1, Complex num2) {
 
 
 /*Given  complex number cmp, and real number "real" return the value of cmp * real*/
-Complex multReal(Complex cmp, float real) {
+Complex multReal(Complex cmp, double real) {
     return getComplex(cmp.real_part * real, cmp.imaginary_part * real);
 }
 
 
 /*Given  complex number cmp, and real number "img"that represent
  * the number img*i return the value of cmp * img * i */
-Complex multImg(Complex cmp, float img) {
+Complex multImg(Complex cmp, double img) {
     return getComplex(-img * cmp.imaginary_part, img * cmp.real_part);
 }
 
@@ -39,7 +39,7 @@ Complex multComp(Complex num1, Complex num2) {
 }
 
 /*Given a complex number num return the value of |num| */
-float comp_abs(Complex num) {
+double comp_abs(Complex num) {
     return sqrt(num.real_part * num.real_part + num.imaginary_part * num.imaginary_part);
 }
